feat: move crm v2 to crm and legacy crm to crm-v1
