+++ src/app/(main)/dashboard/(legacy)/crm-v1/_components/crm.config.ts
