+++ src/app/(main)/dashboard/(legacy)/crm-v1/_components/insight-cards.tsx
