+++ src/app/(main)/dashboard/(legacy)/crm-v1/_components/recent-leads-table/columns.tsx
