--- src/app/(main)/dashboard/crm-v2/page.tsx
@@ -1,15 +1,0 @@
-import { KpiCards } from "./_components/kpi-cards";
-import { OpportunitiesSection } from "./_components/opportunities-section";
-import { PipelineActivity } from "./_components/pipeline-activity";
-import { TaskReminders } from "./_components/task-reminders";
-
-export default function Page() {
-  return (
-    <div className="flex flex-col gap-4 md:gap-6">
-      <KpiCards />
-      <PipelineActivity />
-      <TaskReminders />
-      <OpportunitiesSection />
-    </div>
-  );
-}
