+++ src/app/(main)/dashboard/crm/_components/opportunities-section.tsx
