+++ src/app/(main)/dashboard/crm/_components/opportunities-table/columns.tsx
