+++ src/app/(main)/dashboard/crm/_components/opportunities-table/data.json
