+++ src/app/(main)/dashboard/crm/_components/opportunities-table/schema.ts
