+++ src/app/(main)/dashboard/crm/_components/pipeline-activity.tsx
