+++ src/app/(main)/dashboard/crm/page.tsx
@@ -1,16 +1,15 @@
-import { recentLeadsData } from "./_components/crm.config";
-import { InsightCards } from "./_components/insight-cards";
-import { OperationalCards } from "./_components/operational-cards";
-import { OverviewCards } from "./_components/overview-cards";
-import { RecentLeadsTable } from "./_components/recent-leads-table/table";
+import { KpiCards } from "./_components/kpi-cards";
+import { OpportunitiesSection } from "./_components/opportunities-section";
+import { PipelineActivity } from "./_components/pipeline-activity";
+import { TaskReminders } from "./_components/task-reminders";
 
 export default function Page() {
   return (
     <div className="flex flex-col gap-4 md:gap-6">
-      <OverviewCards />
-      <InsightCards />
-      <OperationalCards />
-      <RecentLeadsTable data={recentLeadsData} />
+      <KpiCards />
+      <PipelineActivity />
+      <TaskReminders />
+      <OpportunitiesSection />
     </div>
   );
 }
